# admin/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Food Delivery Admin APIs",
        "version": "1.0.0",
        "description": "API documentation for the Food Delivery system"
    },
    "servers": [
        {
            "url": "https://0ba3-2401-4900-1c52-2c1d-edce-29a2-737a-21d2.ngrok-free.app/",
            "description": "Dev - Ngrok"
        },
        {
            "url": "http://localhost:3000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/admin/reports/popular-restaurants": {
            "get": {
                "summary": "Get the most popular restaurants",
                "description": "This endpoint allows an admin to retrieve the most popular restaurants based on the number of orders placed.",
                "operationId": "getMostPopularRestaurants",
                "tags": [
                    "Reports"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of popular restaurants",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "restaurantId": {
                                                "type": "integer",
                                                "description": "ID of the restaurant"
                                            },
                                            "orderCount": {
                                                "type": "integer",
                                                "description": "Number of orders placed for this restaurant"
                                            },
                                            "restaurant": {
                                                "type": "object",
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "description": "Name of the restaurant"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/api/admin/reports/order-trends": {
            "get": {
                "summary": "Get order trends by day",
                "description": "This endpoint returns the number of orders placed on each day.",
                "operationId": "getOrderTrends",
                "tags": [
                    "Reports"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order trends by day",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "orderDate": {
                                                "type": "string",
                                                "format": "date",
                                                "description": "The date when the orders were placed"
                                            },
                                            "orderCount": {
                                                "type": "integer",
                                                "description": "Number of orders placed on this date"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT Bearer token for admin authentication. This should be passed in the Authorization header as 'Bearer <your-token>'"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}